# Avoid creating compression map for question only Msg