This avoids creating the compression map, and associated garbage, when the `Msg` only has one question.

~This needs to wait on both #821 & #822, and will probably need to be rebased~.

```
$ benchstat {old,new}.bench
name                       old time/op    new time/op    delta
MsgLength-12                  335ns ± 3%     324ns ± 0%    -3.53%  (p=0.000 n=10+9)
MsgLengthNoCompression-12    9.46ns ± 0%    9.46ns ± 1%      ~     (p=0.623 n=9+9)
MsgLengthPack-12             1.91µs ±31%    2.14µs ±16%      ~     (p=0.219 n=10+9)
MsgLengthOnlyQuestion-12      165ns ± 2%       5ns ± 3%   -96.74%  (p=0.000 n=10+10)
PackMsg-12                   1.59µs ±20%    1.56µs ± 6%      ~     (p=0.905 n=10+9)
PackMsgOnlyQuestion-12        757ns ±33%     232ns ± 4%   -69.28%  (p=0.000 n=10+10)

name                       old alloc/op   new alloc/op   delta
MsgLength-12                  32.0B ± 0%     32.0B ± 0%      ~     (all equal)
MsgLengthNoCompression-12     0.00B          0.00B           ~     (all equal)
MsgLengthPack-12               896B ± 0%      896B ± 0%      ~     (all equal)
MsgLengthOnlyQuestion-12      32.0B ± 0%      0.0B       -100.00%  (p=0.000 n=10+10)
PackMsg-12                     576B ± 0%      576B ± 0%      ~     (all equal)
PackMsgOnlyQuestion-12         320B ± 0%       64B ± 0%   -80.00%  (p=0.000 n=10+10)

name                       old allocs/op  new allocs/op  delta
MsgLength-12                   1.00 ± 0%      1.00 ± 0%      ~     (all equal)
MsgLengthNoCompression-12      0.00           0.00           ~     (all equal)
MsgLengthPack-12               8.00 ± 0%      8.00 ± 0%      ~     (all equal)
MsgLengthOnlyQuestion-12       1.00 ± 0%      0.00       -100.00%  (p=0.000 n=10+10)
PackMsg-12                     7.00 ± 0%      7.00 ± 0%      ~     (all equal)
PackMsgOnlyQuestion-12         3.00 ± 0%      1.00 ± 0%   -66.67%  (p=0.000 n=10+10)
```

Related #821 & #822.
Updates #709.