@tmthrgd 2018-11-27
This is no longer waiting on #821 and can be reviewed/merged.